Removed ClientTorcherino.java (moved to Torcherino.java) Updated to loader 0.4 Updated to 19w14b Moved some more stuff to client only

+++ src/main/java/torcherino/block/screen/StateButtonWidget.java
@@ -42,5 +42,5 @@ public abstract class StateButtonWidget extends ButtonWidget
 	}
 
 	@Override protected String getNarrationMessage(){ return I18n.translate(this.narrationMessage, getStateName(state)); }
-	@Override public void onPress(){ onStateChange(state = (byte) ((state+1) % MAX_STATES)); }
+	@Override public void onPress(){ onStateChange(state =  (state+1) % MAX_STATES); }
 }
